% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.tab.R
\name{anova.tab}
\alias{anova.tab}
\title{Return anova table to a norm way}
\usage{
\method{anova}{tab}(fm)
}
\arguments{
\item{fm:}{the object of aov()}
}
\description{
\code{anova.tab} Return anova table to a norm way
}
\examples{
library(vsnc)
data(MET)
mod <- aov(yield ~ Gen*Loc,data=MET)
summary(mod)
anova.tab(mod)
}
